refactor(web-ui): extract messages handlers from api/mod.rs
Move the four message-related handlers and their supporting types into
a sibling `api/messages.rs`:

- handlers: `send_message`, `send_voice_message`, `quick_message`,
  `stream_run_events`
- request/response types: `SendMessageRequest`, `QuickMessageRequest`,
  `QuickMessageResponse`, `StreamRunEventsQuery`, `VoiceUploadForm`
- SSE event schemas: `SseTokenEvent`, `SseThinkingEvent`,
  `SseStatusEvent`, `SseToolResultEvent`, `SseSubagentStartedEvent`,
  `SseSubagentCompletedEvent`, `SseSubagentTokenEvent`,
  `SseSubagentThinkingEvent`, `SseSubagentToolResultEvent`,
  `SseSubagentStatusEvent`
- 23 handler tests (send_message, send_voice_message, voice_upload,
  stream_run_events, quick_message variants)

Also moves the 5 `stream_conversations_*` tests from `mod.rs` to
`conversations.rs` where they belong (missed in the previous split).

`mod.rs` is now down to ~270 lines holding only `ApiState`, the router,
and the `internal_error` / `empty_string_as_none` / `sse_response`
helpers. `openapi.rs` paths and the router are repointed to the new
locations.

No behavioural changes; `cargo test -p assistant-web-ui --lib` reports
241 passed (same as before). Total lines across the seven `api/*.rs`
files: 4539 → 4570 (+31, accounting for the new module headers and
test-mod imports).
