+++ crates/web-ui/src/api/conversations.rs
@@ -564,6 +564,8 @@ mod tests {
     use uuid::Uuid;
     use wiremock::MockServer;
 
+    use chrono::Utc;
+
     use assistant_storage::ConversationStore;
 
     use super::super::api_router;
@@ -862,5 +864,289 @@ mod tests {
         let arr = json.as_array().unwrap();
         assert_eq!(arr.len(), 1);
         assert_eq!(arr[0]["title"], "Alice conv");
+    }
+    // -- stream_conversations tests --------------------------------------------
+
+    #[tokio::test]
+    async fn stream_conversations_returns_snapshot() {
+        let (state, storage) = event_log_state().await;
+
+        // Create a conversation so the snapshot is non-empty.
+        let store = ConversationStore::for_agent(storage.pool.clone(), "default");
+        store.create_conversation(Some("Hello")).await.unwrap();
+
+        let resp = app(state)
+            .oneshot(
+                Request::builder()
+                    .uri("/conversations/stream")
+                    .body(Body::empty())
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+
+        assert_eq!(resp.status(), StatusCode::OK);
+
+        let body = body_bytes(resp.into_body()).await;
+        let text = String::from_utf8_lossy(&body);
+
+        // Should contain a snapshot event.
+        assert!(
+            text.contains("event: snapshot"),
+            "response should contain snapshot event, got: {text}"
+        );
+        assert!(
+            text.contains("Hello"),
+            "snapshot should contain the conversation title"
+        );
+    }
+
+    #[tokio::test]
+    async fn stream_conversations_forwards_upserted_delta() {
+        let (state, _storage) = event_log_state().await;
+        let broadcaster = state.conversation_broadcaster.clone();
+        let pool = state.pool.clone();
+
+        // Spawn the SSE stream reader in a task.
+        let app = app(state);
+        let handle = tokio::spawn(async move {
+            let resp = app
+                .oneshot(
+                    Request::builder()
+                        .uri("/conversations/stream")
+                        .body(Body::empty())
+                        .unwrap(),
+                )
+                .await
+                .unwrap();
+            assert_eq!(resp.status(), StatusCode::OK);
+            body_bytes(resp.into_body()).await
+        });
+
+        // Give the stream a moment to subscribe.
+        tokio::time::sleep(std::time::Duration::from_millis(50)).await;
+
+        // Create a conversation through a broadcaster-wired store.
+        let store = ConversationStore::for_agent(pool, "default").with_broadcaster(broadcaster);
+        store.create_conversation(Some("Delta Test")).await.unwrap();
+
+        // Give the stream a moment to forward the event, then drop broadcaster
+        // to close the channel so the SSE task finishes.
+        tokio::time::sleep(std::time::Duration::from_millis(100)).await;
+        drop(store);
+
+        let body = tokio::time::timeout(std::time::Duration::from_secs(5), handle)
+            .await
+            .expect("stream should finish within timeout")
+            .expect("stream task should not panic");
+        let text = String::from_utf8_lossy(&body);
+
+        assert!(
+            text.contains("event: upserted"),
+            "stream should contain upserted event, got: {text}"
+        );
+        assert!(
+            text.contains("Delta Test"),
+            "upserted event should contain the conversation title"
+        );
+    }
+
+    #[tokio::test]
+    async fn test_update_title_broadcasts_upserted_with_new_title() {
+        // Closes the loop the title-generator worker depends on: when the
+        // worker calls update_title on an existing conversation, an Upserted
+        // event MUST be broadcast carrying the new title so any subscribed
+        // UI re-renders without a refresh.  The wire-level forward to SSE
+        // is covered by `stream_conversations_forwards_upserted_delta`.
+        let (state, _storage) = event_log_state().await;
+        let broadcaster = state.conversation_broadcaster.clone();
+        let pool = state.pool.clone();
+
+        let store =
+            ConversationStore::for_agent(pool, "default").with_broadcaster(broadcaster.clone());
+        let conv = store.create_conversation(None).await.unwrap();
+        let conv_id = conv.id;
+
+        // Subscribe AFTER create so we only capture the update.
+        let mut rx = broadcaster.subscribe();
+        store
+            .update_title(conv_id, "Generated By Worker")
+            .await
+            .unwrap();
+
+        let event = tokio::time::timeout(std::time::Duration::from_secs(2), rx.recv())
+            .await
+            .expect("broadcast must arrive within 2s")
+            .expect("recv must not error");
+
+        match event {
+            assistant_storage::ConversationEvent::Upserted(record) => {
+                assert_eq!(record.id, conv_id);
+                assert_eq!(record.title.as_deref(), Some("Generated By Worker"));
+                assert!(
+                    record.title_locked,
+                    "broadcast after update_title must carry title_locked = true",
+                );
+            }
+            other => panic!("expected Upserted, got {other:?}"),
+        }
+    }
+
+    #[tokio::test]
+    async fn stream_conversations_filters_by_agent_id() {
+        use assistant_storage::ConversationEvent;
+
+        let (state, _storage) = event_log_state().await;
+        let broadcaster = state.conversation_broadcaster.clone();
+
+        // Stream filtered to agent "work".
+        let app = app(state);
+        let handle = tokio::spawn(async move {
+            let resp = app
+                .oneshot(
+                    Request::builder()
+                        .uri("/conversations/stream?agent_id=work")
+                        .body(Body::empty())
+                        .unwrap(),
+                )
+                .await
+                .unwrap();
+            body_bytes(resp.into_body()).await
+        });
+
+        tokio::time::sleep(std::time::Duration::from_millis(50)).await;
+
+        // Emit an event for agent "default" — should be filtered out.
+        broadcaster.emit(ConversationEvent::Upserted(
+            assistant_storage::ConversationRecord {
+                id: Uuid::new_v4(),
+                agent_id: "default".to_string(),
+                title: Some("DefaultConv".to_string()),
+                title_locked: true,
+                user_id: None,
+                created_at: Utc::now(),
+                updated_at: Utc::now(),
+            },
+        ));
+
+        // Emit an event for agent "work" — should pass through.
+        broadcaster.emit(ConversationEvent::Upserted(
+            assistant_storage::ConversationRecord {
+                id: Uuid::new_v4(),
+                agent_id: "work".to_string(),
+                title: Some("WorkConv".to_string()),
+                title_locked: true,
+                user_id: None,
+                created_at: Utc::now(),
+                updated_at: Utc::now(),
+            },
+        ));
+
+        tokio::time::sleep(std::time::Duration::from_millis(100)).await;
+        drop(broadcaster);
+
+        let body = tokio::time::timeout(std::time::Duration::from_secs(5), handle)
+            .await
+            .expect("timeout")
+            .expect("no panic");
+        let text = String::from_utf8_lossy(&body);
+
+        assert!(
+            !text.contains("DefaultConv"),
+            "should NOT contain event for wrong agent, got: {text}"
+        );
+        assert!(
+            text.contains("WorkConv"),
+            "should contain event for filtered agent, got: {text}"
+        );
+    }
+
+    #[tokio::test]
+    async fn stream_conversations_e2e_create_and_delete() {
+        let (state, _storage) = event_log_state().await;
+
+        // Clone app router so we can issue multiple requests.
+        let router = app(state);
+
+        // Spawn the SSE stream reader in a background task.
+        let stream_router = router.clone();
+        let handle = tokio::spawn(async move {
+            let resp = stream_router
+                .oneshot(
+                    Request::builder()
+                        .uri("/conversations/stream")
+                        .body(Body::empty())
+                        .unwrap(),
+                )
+                .await
+                .unwrap();
+            assert_eq!(resp.status(), StatusCode::OK);
+            body_bytes(resp.into_body()).await
+        });
+
+        // Give the stream time to subscribe.
+        tokio::time::sleep(std::time::Duration::from_millis(50)).await;
+
+        // Create a conversation via REST.
+        let create_resp = router
+            .clone()
+            .oneshot(
+                Request::builder()
+                    .method("POST")
+                    .uri("/conversations")
+                    .header("Content-Type", "application/json")
+                    .body(Body::from(r#"{"title":"E2E Conv"}"#))
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+        assert_eq!(create_resp.status(), StatusCode::CREATED);
+        let json = body_json(create_resp.into_body()).await;
+        let conv_id = json["id"].as_str().expect("should have an id").to_string();
+
+        // Give the stream time to forward the upserted event.
+        tokio::time::sleep(std::time::Duration::from_millis(50)).await;
+
+        // Delete the conversation via REST.
+        let delete_resp = router
+            .clone()
+            .oneshot(
+                Request::builder()
+                    .method("DELETE")
+                    .uri(format!("/conversations/{conv_id}"))
+                    .body(Body::empty())
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+        assert_eq!(delete_resp.status(), StatusCode::NO_CONTENT);
+
+        // Give the stream time to forward the deleted event, then drop
+        // broadcaster indirectly by dropping the router.
+        tokio::time::sleep(std::time::Duration::from_millis(100)).await;
+        drop(router);
+
+        let body = tokio::time::timeout(std::time::Duration::from_secs(5), handle)
+            .await
+            .expect("stream should finish within timeout")
+            .expect("stream task should not panic");
+        let text = String::from_utf8_lossy(&body);
+
+        assert!(
+            text.contains("event: upserted"),
+            "stream should contain upserted event from REST create, got: {text}"
+        );
+        assert!(
+            text.contains("E2E Conv"),
+            "upserted event should contain conversation title"
+        );
+        assert!(
+            text.contains("event: deleted"),
+            "stream should contain deleted event from REST delete, got: {text}"
+        );
+        assert!(
+            text.contains(&conv_id),
+            "deleted event should contain the conversation ID"
+        );
     }
 }
