+++ crates/web-ui/src/api/mod.rs
@@ -21,6 +21,7 @@ pub mod conversations;
 pub mod interfaces;
 pub mod logs;
 pub mod members;
+pub mod messages;
 pub mod orgs;
 pub mod personas;
 pub mod push;
@@ -103,34 +104,30 @@ pub(crate) fn sse_response(rx: mpsc::Receiver<Result<Event, Infallible>>) -> Res
 // | PATCH    | `/api/conversations/{id}`             | Update conversation title       |
 // | POST     | `/api/conversations/{id}/messages`    | Send a message (SSE stream)     |
 
+use std::collections::HashMap;
 use std::convert::Infallible;
 use std::sync::Arc;
 
-use assistant_transcription::{TranscriptionProvider, TranscriptionRequest, TtsProvider};
-
-use std::collections::HashMap;
-
-use assistant_core::{ConversationConfig, types::conversation::Interface};
-use assistant_runtime::{AssistantInterface, CommandRegistry, Orchestrator, OrchestratorEvent};
+use assistant_core::ConversationConfig;
+use assistant_runtime::{AssistantInterface, CommandRegistry, Orchestrator};
 use assistant_storage::{
     AttachmentStore, CommandEventStore, ConversationBroadcast, ConversationEventStore,
-    ConversationStore, InMemoryConversationBroadcaster, RunBroadcaster,
+    InMemoryConversationBroadcaster, RunBroadcaster,
 };
+use assistant_transcription::{TranscriptionProvider, TtsProvider};
 use axum::{
-    Json, Router,
-    extract::{DefaultBodyLimit, Multipart, Path, State},
-    http::{StatusCode, header},
+    Router,
+    extract::DefaultBodyLimit,
+    http::header,
     response::{
         IntoResponse, Response,
         sse::{Event, KeepAlive, Sse},
     },
     routing::{delete, get, patch, post},
 };
-use serde::{Deserialize, Serialize};
 use sqlx::SqlitePool;
 use tokio::sync::{RwLock, mpsc};
 use tokio_stream::wrappers::ReceiverStream;
-use tracing::warn;
 use uuid::Uuid;
 
 // -- State -------------------------------------------------------------------
@@ -220,44 +217,6 @@ impl ApiState {
     }
 }
 
-/// Body for `POST /api/conversations/{id}/messages`.
-#[derive(Debug, Deserialize, utoipa::ToSchema)]
-#[schema(as = ApiSendMessageRequest)]
-pub struct SendMessageRequest {
-    /// The message text to send to the assistant.
-    pub message: String,
-    /// Optional attachment IDs to include with this message.
-    #[serde(default)]
-    pub attachment_ids: Vec<Uuid>,
-}
-
-/// Body for `POST /api/quick-message`.
-#[derive(Debug, Deserialize, utoipa::ToSchema)]
-pub struct QuickMessageRequest {
-    /// The message text to send to the assistant.
-    pub message: String,
-    /// Optional persona ID to route the message to a specific persona.
-    /// When absent, the server's active persona is used.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub persona_id: Option<String>,
-    /// Optional context text to include with the message (e.g. clipboard
-    /// contents, file text).  Prepended to the user message for the LLM.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub context: Option<String>,
-}
-
-/// Response for `POST /api/quick-message`.
-#[derive(Debug, Serialize, utoipa::ToSchema)]
-pub struct QuickMessageResponse {
-    /// The ID of the newly created conversation.
-    pub conversation_id: Uuid,
-    /// The ID of the persisted assistant message, if available.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub message_id: Option<Uuid>,
-    /// The full assistant response text.
-    pub answer: String,
-}
-
 // -- Router ------------------------------------------------------------------
 
 /// Build the conversations API sub-router.  Mounted under `/api`.
@@ -279,9 +238,12 @@ pub fn api_router() -> Router<ApiState> {
             "/conversations/{id}",
             patch(conversations::update_conversation),
         )
-        .route("/conversations/{id}/messages", post(send_message))
-        .route("/conversations/{id}/voice", post(send_voice_message))
-        .route("/quick-message", post(quick_message))
+        .route("/conversations/{id}/messages", post(messages::send_message))
+        .route(
+            "/conversations/{id}/voice",
+            post(messages::send_voice_message),
+        )
+        .route("/quick-message", post(messages::quick_message))
         .route(
             "/conversations/{id}/attachments",
             post(attachments::upload_attachment).layer(DefaultBodyLimit::max(27 * 1024 * 1024)), // 27 MB
@@ -289,7 +251,7 @@ pub fn api_router() -> Router<ApiState> {
         .route("/attachments/{id}", get(attachments::serve_attachment))
         .route(
             "/conversations/{id}/runs/{run_id}/events/stream",
-            get(stream_run_events),
+            get(messages::stream_run_events),
         )
         .route("/messages/{id}/audio", get(audio::get_message_audio))
         .route("/audio/{id}", get(audio::get_audio))
@@ -305,2344 +267,3 @@ pub fn api_router() -> Router<ApiState> {
 }
 
 // -- Handlers ----------------------------------------------------------------
-
-/// `POST /api/conversations/{id}/messages` — send a message and stream the response.
-///
-/// The response is a `text/event-stream` (SSE) with two event types:
-/// - `event: token` — incremental assistant token (data is plain text)
-/// - `event: done`  — final JSON object: `{"role":"assistant","content":"..."}`
-#[utoipa::path(
-    post,
-    path = "/api/conversations/{id}/messages",
-    tag = "conversations",
-    params(("id" = Uuid, Path, description = "Conversation ID")),
-    request_body = SendMessageRequest,
-    responses(
-        (status = 200, description = "SSE stream of assistant tokens and final message",
-         content_type = "text/event-stream"),
-        (status = 400, description = "Invalid ID or empty message"),
-        (status = 401, description = "Unauthorized"),
-        (status = 404, description = "Conversation not found"),
-    ),
-    security(("bearer_token" = []))
-)]
-pub async fn send_message(
-    State(state): State<ApiState>,
-    Path(id): Path<String>,
-    Json(body): Json<SendMessageRequest>,
-) -> Response {
-    let conv_id = match Uuid::parse_str(&id) {
-        Ok(u) => u,
-        Err(_) => return (StatusCode::BAD_REQUEST, "Invalid conversation ID").into_response(),
-    };
-
-    let content = body.message.trim().to_string();
-    if content.is_empty() {
-        return (StatusCode::BAD_REQUEST, "Message cannot be empty").into_response();
-    }
-
-    // Verify the conversation exists before streaming.
-    let agent_id = state.agent_id.read().await.clone();
-    let store = ConversationStore::for_agent(state.pool.clone(), &agent_id)
-        .with_broadcaster(state.conversation_broadcaster.clone());
-
-    match store.get_conversation(conv_id).await {
-        Ok(None) => {
-            return (StatusCode::NOT_FOUND, "Conversation not found").into_response();
-        }
-        Err(e) => {
-            warn!("Failed to check conversation {conv_id}: {e}");
-            return (StatusCode::INTERNAL_SERVER_ERROR, "Database error").into_response();
-        }
-        _ => {}
-    }
-
-    // Title generation is owned by the title-generator worker (consumer of
-    // `turn.result`), not this handler. See `crates/runtime/src/title_generator.rs`.
-
-    // Generate a unique ID for this orchestrator run.
-    let run_id = Uuid::new_v4();
-
-    let (sse_tx, sse_rx) = mpsc::channel::<Result<Event, Infallible>>(64);
-    let (event_tx, mut event_rx) = mpsc::channel::<OrchestratorEvent>(64);
-
-    state
-        .orchestrator
-        .register_token_sink(conv_id, event_tx)
-        .await;
-
-    let orchestrator = state.orchestrator.clone();
-    let user_attachment_ids = body.attachment_ids;
-    let turn_result_rx = {
-        let (tx, rx) =
-            tokio::sync::oneshot::channel::<anyhow::Result<assistant_runtime::TurnResult>>();
-        tokio::spawn(async move {
-            let result = if user_attachment_ids.is_empty() {
-                orchestrator
-                    .submit_turn(&content, conv_id, Interface::Web, None)
-                    .await
-            } else {
-                orchestrator
-                    .submit_turn_with_attachments(
-                        &content,
-                        conv_id,
-                        Interface::Web,
-                        None,
-                        user_attachment_ids,
-                    )
-                    .await
-            };
-            let _ = tx.send(result);
-        });
-        rx
-    };
-
-    let push_dispatcher_for_sse = state.push_dispatcher.clone();
-    let conv_id_for_push = conv_id;
-    let event_store = state.event_store.clone();
-    let run_broadcaster = state.run_broadcaster.clone();
-
-    // Emit run_started event (sequence 0) to the durable log.
-    let run_id_str = run_id.to_string();
-    let run_started_payload = serde_json::json!({"run_id": run_id_str});
-    if let Err(e) = event_store
-        .append_event(
-            &run_id_str,
-            &conv_id.to_string(),
-            0,
-            "run_started",
-            &run_started_payload,
-        )
-        .await
-    {
-        warn!("Failed to persist run_started event for run {run_id}: {e}");
-    }
-
-    // Register broadcast channel for live tailing.
-    let broadcast_tx = run_broadcaster.start_run(run_id).await;
-
-    // Emit run_started as first SSE event so the client learns the run_id.
-    let run_started_sse = Event::default()
-        .event("run_started")
-        .data(run_started_payload.to_string())
-        .id("0");
-    if sse_tx.send(Ok(run_started_sse)).await.is_err() {
-        run_broadcaster.finish_run(&run_id).await;
-        return sse_response(sse_rx);
-    }
-
-    tokio::spawn(async move {
-        let mut full_text = String::new();
-        let mut seq: i64 = 1; // sequence 0 was run_started
-        let conv_id_str = conv_id_for_push.to_string();
-
-        // Throttle thinking persistence: accumulate tokens, flush every 20 tokens
-        // or when a non-thinking event arrives. SSE is still sent immediately.
-        let mut thinking_buf = String::new();
-        let mut thinking_token_count: usize = 0;
-        const THINKING_BATCH_SIZE: usize = 20;
-
-        while let Some(orch_event) = event_rx.recv().await {
-            // Flush accumulated thinking buffer before non-thinking events.
-            let is_thinking = matches!(orch_event, OrchestratorEvent::Thinking(_));
-            if !is_thinking && !thinking_buf.is_empty() {
-                let p = serde_json::json!({"content": thinking_buf});
-                if let Err(e) = event_store
-                    .append_event(&run_id_str, &conv_id_str, seq, "thinking", &p)
-                    .await
-                {
-                    warn!("Failed to persist thinking batch seq={seq} for run {run_id}: {e}");
-                }
-                let live = assistant_storage::LiveEvent {
-                    sequence: seq,
-                    event_type: "thinking".to_string(),
-                    payload: p,
-                };
-                let _ = broadcast_tx.send(live);
-                seq += 1;
-                thinking_buf.clear();
-                thinking_token_count = 0;
-            }
-
-            let (event_type, payload, sse_event) = match orch_event {
-                OrchestratorEvent::Token(ref token) => {
-                    full_text.push_str(token);
-                    let p = serde_json::json!({"token": token});
-                    let e = Event::default().event("token").data(token.clone());
-                    ("token", p, e)
-                }
-                OrchestratorEvent::Status {
-                    ref message,
-                    ref tool_call_id,
-                } => {
-                    let mut p = serde_json::json!({"message": message});
-                    if let Some(id) = tool_call_id {
-                        p["tool_call_id"] = serde_json::Value::String(id.clone());
-                    }
-                    let e = Event::default().event("status").data(p.to_string());
-                    ("status", p, e)
-                }
-                OrchestratorEvent::ToolResult {
-                    ref tool_name,
-                    ref status,
-                    ref arguments,
-                    ref result,
-                    ref tool_call_id,
-                } => {
-                    let mut p = serde_json::json!({"tool_name": tool_name, "status": status});
-                    if let Some(args) = arguments {
-                        p["arguments"] = args.clone();
-                    }
-                    if let Some(res) = result {
-                        p["result"] = serde_json::Value::String(res.clone());
-                    }
-                    if let Some(id) = tool_call_id {
-                        p["tool_call_id"] = serde_json::Value::String(id.clone());
-                    }
-                    let e = Event::default().event("tool_result").data(p.to_string());
-                    ("tool_result", p, e)
-                }
-                OrchestratorEvent::SkillComplete {
-                    ref skill_name,
-                    success,
-                    ref summary,
-                } => {
-                    if let Some(ref dispatcher) = push_dispatcher_for_sse {
-                        let title = if success {
-                            "Skill complete"
-                        } else {
-                            "Skill failed"
-                        };
-                        let body = format!("{skill_name}: {summary}");
-                        let cid = conv_id_str.clone();
-                        let d = dispatcher.clone();
-                        tokio::spawn(async move {
-                            if let Err(e) = d.send_to_all(title, &body, Some(&cid)).await {
-                                warn!("Push (skill) failed: {e}");
-                            }
-                        });
-                    }
-                    let p = serde_json::json!({"skill_name": skill_name, "success": success, "summary": summary});
-                    let e = Event::default().event("skill_complete").data(p.to_string());
-                    ("skill_complete", p, e)
-                }
-                OrchestratorEvent::AgentError { ref message } => {
-                    if let Some(ref dispatcher) = push_dispatcher_for_sse {
-                        let body = message.chars().take(80).collect::<String>();
-                        let cid = conv_id_str.clone();
-                        let d = dispatcher.clone();
-                        tokio::spawn(async move {
-                            if let Err(e) =
-                                d.send_to_all("Assistant error", &body, Some(&cid)).await
-                            {
-                                warn!("Push (agent error) failed: {e}");
-                            }
-                        });
-                    }
-                    let p = serde_json::json!({"message": message});
-                    let e = Event::default().event("agent_error").data(message.clone());
-                    ("agent_error", p, e)
-                }
-                OrchestratorEvent::Thinking(ref content) => {
-                    // Send SSE immediately for real-time rendering.
-                    let p = serde_json::json!({"content": content});
-                    let e = Event::default().event("thinking").data(p.to_string());
-                    if sse_tx.send(Ok(e)).await.is_err() {
-                        break;
-                    }
-                    // Accumulate for batched persistence.
-                    thinking_buf.push_str(content);
-                    thinking_token_count += 1;
-                    if thinking_token_count >= THINKING_BATCH_SIZE {
-                        let batch_p = serde_json::json!({"content": thinking_buf});
-                        if let Err(e) = event_store
-                            .append_event(&run_id_str, &conv_id_str, seq, "thinking", &batch_p)
-                            .await
-                        {
-                            warn!(
-                                "Failed to persist thinking batch seq={seq} for run {run_id}: {e}"
-                            );
-                        }
-                        let live = assistant_storage::LiveEvent {
-                            sequence: seq,
-                            event_type: "thinking".to_string(),
-                            payload: batch_p,
-                        };
-                        let _ = broadcast_tx.send(live);
-                        seq += 1;
-                        thinking_buf.clear();
-                        thinking_token_count = 0;
-                    }
-                    continue;
-                }
-                OrchestratorEvent::SubagentStarted {
-                    ref agent_id,
-                    ref task,
-                } => {
-                    let p = serde_json::json!({"agent_id": agent_id, "task": task});
-                    let e = Event::default()
-                        .event("subagent_started")
-                        .data(p.to_string());
-                    ("subagent_started", p, e)
-                }
-                OrchestratorEvent::SubagentCompleted {
-                    ref agent_id,
-                    ref status,
-                    ref summary,
-                } => {
-                    let p = serde_json::json!({"agent_id": agent_id, "status": status, "summary": summary});
-                    let e = Event::default()
-                        .event("subagent_completed")
-                        .data(p.to_string());
-                    ("subagent_completed", p, e)
-                }
-                OrchestratorEvent::SubagentEvent {
-                    ref agent_id,
-                    ref inner,
-                } => {
-                    let (event_name, inner_data) = match inner.as_ref() {
-                        OrchestratorEvent::Token(t) => {
-                            ("subagent_token", serde_json::json!({"content": t}))
-                        }
-                        OrchestratorEvent::Thinking(t) => {
-                            ("subagent_thinking", serde_json::json!({"content": t}))
-                        }
-                        OrchestratorEvent::Status {
-                            message,
-                            tool_call_id,
-                        } => {
-                            let mut d = serde_json::json!({"message": message});
-                            if let Some(id) = tool_call_id {
-                                d["tool_call_id"] = serde_json::Value::String(id.clone());
-                            }
-                            ("subagent_status", d)
-                        }
-                        OrchestratorEvent::ToolResult {
-                            tool_name,
-                            status,
-                            arguments,
-                            result,
-                            tool_call_id,
-                        } => {
-                            let mut d = serde_json::json!({"tool_name": tool_name, "status": status, "arguments": arguments, "result": result});
-                            if let Some(id) = tool_call_id {
-                                d["tool_call_id"] = serde_json::Value::String(id.clone());
-                            }
-                            ("subagent_tool_result", d)
-                        }
-                        other => (
-                            "subagent_event",
-                            serde_json::json!({"type": format!("{other:?}")}),
-                        ),
-                    };
-                    let p = serde_json::json!({"agent_id": agent_id, "data": inner_data});
-                    let e = Event::default().event(event_name).data(p.to_string());
-                    (event_name, p, e)
-                }
-                OrchestratorEvent::AudioReady { ref audio_id } => {
-                    let p = serde_json::json!({"audio_id": audio_id, "auto_play": true});
-                    let e = Event::default().event("audio_ready").data(p.to_string());
-                    ("audio_ready", p, e)
-                }
-            };
-
-            // Persist to durable log.
-            if let Err(e) = event_store
-                .append_event(&run_id_str, &conv_id_str, seq, event_type, &payload)
-                .await
-            {
-                warn!("Failed to persist event seq={seq} for run {run_id}: {e}");
-            }
-            // Broadcast to live tail subscribers.
-            let live = assistant_storage::LiveEvent {
-                sequence: seq,
-                event_type: event_type.to_string(),
-                payload: payload.clone(),
-            };
-            let _ = broadcast_tx.send(live);
-            let sse_event = sse_event.id(seq.to_string());
-            seq += 1;
-
-            if sse_tx.send(Ok(sse_event)).await.is_err() {
-                break;
-            }
-        }
-
-        // Flush any remaining thinking tokens that didn't reach the batch threshold.
-        if !thinking_buf.is_empty() {
-            let p = serde_json::json!({"content": thinking_buf});
-            if let Err(e) = event_store
-                .append_event(&run_id_str, &conv_id_str, seq, "thinking", &p)
-                .await
-            {
-                warn!("Failed to persist final thinking batch seq={seq} for run {run_id}: {e}");
-            }
-            let live = assistant_storage::LiveEvent {
-                sequence: seq,
-                event_type: "thinking".to_string(),
-                payload: p,
-            };
-            let _ = broadcast_tx.send(live);
-            seq += 1;
-        }
-
-        let (reply_text, reply_message_id, reply_attachment_ids) = match turn_result_rx.await {
-            Ok(Ok(result)) => (result.answer, result.message_id, result.attachment_ids),
-            Ok(Err(_)) | Err(_) => (full_text, None, vec![]),
-        };
-
-        let mut done_data = serde_json::json!({
-            "role": "assistant",
-            "content": reply_text,
-        });
-        if let Some(mid) = reply_message_id {
-            done_data["message_id"] = serde_json::Value::String(mid.to_string());
-        }
-        if !reply_attachment_ids.is_empty() {
-            done_data["attachment_ids"] = serde_json::json!(
-                reply_attachment_ids
-                    .iter()
-                    .map(|id| id.to_string())
-                    .collect::<Vec<_>>()
-            );
-        }
-
-        // Persist done event.
-        if let Err(e) = event_store
-            .append_event(&run_id_str, &conv_id_str, seq, "done", &done_data)
-            .await
-        {
-            warn!("Failed to persist done event for run {run_id}: {e}");
-        }
-        let _ = broadcast_tx.send(assistant_storage::LiveEvent {
-            sequence: seq,
-            event_type: "done".to_string(),
-            payload: done_data.clone(),
-        });
-
-        let done = Event::default()
-            .event("done")
-            .data(done_data.to_string())
-            .id(seq.to_string());
-        let _ = sse_tx.send(Ok(done)).await;
-
-        // Signal run completion — drops broadcast channel, subscribers observe close.
-        run_broadcaster.finish_run(&run_id).await;
-
-        // Fire Web Push notification after emitting the SSE done event.
-        if let Some(dispatcher) = push_dispatcher_for_sse {
-            let body = reply_text.chars().take(80).collect::<String>();
-            if let Err(e) = dispatcher
-                .send_to_all("New message", &body, Some(&conv_id_str))
-                .await
-            {
-                warn!("Push dispatch failed: {e}");
-            }
-        }
-    });
-
-    // Include run_id in response headers as a fallback for clients that crash
-    // before receiving the run_started SSE event.
-    let mut response = sse_response(sse_rx);
-    if let Ok(hv) = run_id.to_string().parse::<header::HeaderValue>() {
-        response.headers_mut().insert("X-Run-Id", hv);
-    }
-    response
-}
-
-// -- Run event replay / tail -------------------------------------------------
-
-#[derive(serde::Deserialize, utoipa::ToSchema)]
-pub struct StreamRunEventsQuery {
-    /// Replay from this sequence number (inclusive). Defaults to 0.
-    pub since: Option<i64>,
-}
-
-// -- SSE event payload schemas ------------------------------------------------
-
-/// SSE `token` event — a text chunk from the assistant's response.
-#[derive(serde::Serialize, utoipa::ToSchema)]
-pub struct SseTokenEvent {
-    /// The text content of this token chunk.
-    pub content: String,
-}
-
-/// SSE `thinking` event — a thinking/reasoning token from the model.
-#[derive(serde::Serialize, utoipa::ToSchema)]
-pub struct SseThinkingEvent {
-    /// The thinking text content.
-    pub content: String,
-}
-
-/// SSE `status` event — a status message (e.g. "Calling tool X").
-#[derive(serde::Serialize, utoipa::ToSchema)]
-pub struct SseStatusEvent {
-    /// Human-readable status message.
-    pub message: String,
-}
-
-/// SSE `tool_result` event — result of a tool invocation.
-#[derive(serde::Serialize, utoipa::ToSchema)]
-pub struct SseToolResultEvent {
-    /// Name of the tool that was called.
-    pub tool_name: String,
-    /// Whether the tool call succeeded or failed.
-    pub status: String,
-    /// The arguments passed to the tool (JSON string).
-    pub arguments: Option<String>,
-    /// The tool's output or error message.
-    pub result: Option<String>,
-}
-
-/// SSE `subagent_started` event — a subagent has been spawned.
-#[derive(serde::Serialize, utoipa::ToSchema)]
-pub struct SseSubagentStartedEvent {
-    /// Unique identifier of the subagent.
-    pub agent_id: String,
-    /// The task description given to the subagent.
-    pub task: String,
-}
-
-/// SSE `subagent_completed` event — a subagent has finished.
-#[derive(serde::Serialize, utoipa::ToSchema)]
-pub struct SseSubagentCompletedEvent {
-    /// Unique identifier of the subagent.
-    pub agent_id: String,
-    /// Completion status (e.g. "success", "error").
-    pub status: String,
-    /// Summary of the subagent's work.
-    pub summary: Option<String>,
-}
-
-/// SSE `subagent_token` event — a text token from a subagent's response.
-#[derive(serde::Serialize, utoipa::ToSchema)]
-pub struct SseSubagentTokenEvent {
-    /// The subagent that produced this token.
-    pub agent_id: String,
-    /// Inner event payload.
-    pub data: SseTokenEvent,
-}
-
-/// SSE `subagent_thinking` event — a thinking token from a subagent.
-#[derive(serde::Serialize, utoipa::ToSchema)]
-pub struct SseSubagentThinkingEvent {
-    /// The subagent that produced this thinking token.
-    pub agent_id: String,
-    /// Inner event payload.
-    pub data: SseThinkingEvent,
-}
-
-/// SSE `subagent_tool_result` event — a tool result from a subagent.
-#[derive(serde::Serialize, utoipa::ToSchema)]
-pub struct SseSubagentToolResultEvent {
-    /// The subagent that executed the tool.
-    pub agent_id: String,
-    /// Inner event payload.
-    pub data: SseToolResultEvent,
-}
-
-/// SSE `subagent_status` event — a status update from a subagent.
-#[derive(serde::Serialize, utoipa::ToSchema)]
-pub struct SseSubagentStatusEvent {
-    /// The subagent that emitted this status.
-    pub agent_id: String,
-    /// Inner event payload.
-    pub data: SseStatusEvent,
-}
-
-/// `GET /api/conversations/{id}/runs/{run_id}/events/stream`
-///
-/// Replays stored events from `?since` (default 0), then tails live events
-/// if the run is still active.  Closes automatically when the `done` or
-/// `error` event is reached.
-///
-/// Returns:
-/// - `404` if no events exist for `run_id` (run never started or unknown)
-/// - `410` if the run existed but all events have been pruned (TTL elapsed)
-#[utoipa::path(
-    get,
-    path = "/api/conversations/{id}/runs/{run_id}/events/stream",
-    tag = "conversations",
-    params(
-        ("id" = String, Path, description = "Conversation UUID"),
-        ("run_id" = String, Path, description = "Run UUID from run_started event"),
-        ("since" = Option<i64>, Query, description = "Replay from this sequence number (default 0)"),
-    ),
-    responses(
-        (status = 200, description = "SSE stream of run events", content_type = "text/event-stream"),
-        (status = 404, description = "Run not found"),
-        (status = 410, description = "Run events expired"),
-    ),
-    security(("bearer_token" = []))
-)]
-pub async fn stream_run_events(
-    State(state): State<ApiState>,
-    Path((conv_id_str, run_id_str)): Path<(String, String)>,
-    axum::extract::Query(query): axum::extract::Query<StreamRunEventsQuery>,
-) -> Response {
-    let run_id = match Uuid::parse_str(&run_id_str) {
-        Ok(u) => u,
-        Err(_) => return (StatusCode::BAD_REQUEST, "Invalid run ID").into_response(),
-    };
-    let since = query.since.unwrap_or(0);
-
-    // Check whether any events exist for this run.
-    let event_store = &state.event_store;
-    match event_store.has_events(&run_id_str).await {
-        Ok(false) => {
-            // Could be: run never existed, or all events were pruned.
-            // Distinguish by checking if the run is active in the broadcaster.
-            if state.run_broadcaster.subscribe(&run_id).await.is_none() {
-                // Not in the live registry either — 404.
-                return (
-                    StatusCode::NOT_FOUND,
-                    axum::Json(serde_json::json!({"error": "run not found"})),
-                )
-                    .into_response();
-            }
-            // It's active but has no DB events yet (race condition on startup).
-            // Fall through to live-tail only.
-        }
-        Ok(true) => {
-            // Check if it was pruned (no events AND no live registration).
-            // has_events returned true, so there are still rows — continue.
-        }
-        Err(e) => {
-            warn!("event store error for run {run_id}: {e}");
-            return (StatusCode::INTERNAL_SERVER_ERROR, "Database error").into_response();
-        }
-    }
-
-    // Check if events were pruned (run existed, events gone).
-    // A run_id with zero rows that is NOT in the broadcaster means it's 410.
-    // (Handled above — if has_events is false and no live entry → 404.)
-    // For simplicity we treat unknown old runs as 404.
-
-    let (sse_tx, sse_rx) = mpsc::channel::<Result<Event, Infallible>>(64);
-
-    // Subscribe to live broadcast channel before replaying, so we don't miss
-    // any events emitted between the DB read and the subscription.
-    let live_rx = state.run_broadcaster.subscribe(&run_id).await;
-    let is_active = live_rx.is_some();
-
-    // Replay stored events.
-    let past_events = match event_store.list_events_since(&run_id_str, since).await {
-        Ok(v) => v,
-        Err(e) => {
-            warn!("Failed to list events for run {run_id}: {e}");
-            return (StatusCode::INTERNAL_SERVER_ERROR, "Database error").into_response();
-        }
-    };
-
-    // Check if the run completed before we even arrived.
-    let already_complete = past_events
-        .iter()
-        .any(|e| e.event_type == "done" || e.event_type == "agent_error");
-
-    let conv_id_str_clone = conv_id_str.clone();
-    tokio::spawn(async move {
-        // Send all replayed events.
-        for row in past_events {
-            let ev = Event::default()
-                .event(row.event_type.as_str())
-                .data(row.payload.to_string())
-                .id(row.sequence.to_string());
-            if sse_tx.send(Ok(ev)).await.is_err() {
-                return;
-            }
-        }
-
-        // If the run was already complete, close the stream.
-        if already_complete || !is_active {
-            return;
-        }
-
-        // Tail live events.
-        let mut live_rx = match live_rx {
-            Some(r) => r,
-            None => return,
-        };
-
-        loop {
-            match live_rx.recv().await {
-                Ok(live) => {
-                    let ev = Event::default()
-                        .event(live.event_type.as_str())
-                        .data(live.payload.to_string())
-                        .id(live.sequence.to_string());
-                    let is_terminal = live.event_type == "done" || live.event_type == "agent_error";
-                    if sse_tx.send(Ok(ev)).await.is_err() {
-                        return;
-                    }
-                    if is_terminal {
-                        return;
-                    }
-                }
-                Err(tokio::sync::broadcast::error::RecvError::Closed) => {
-                    // Run completed and sender was dropped.
-                    return;
-                }
-                Err(tokio::sync::broadcast::error::RecvError::Lagged(n)) => {
-                    warn!(
-                        "stream_run_events lagged by {n} events for run {run_id} conv {conv_id_str_clone}"
-                    );
-                    // Continue — we already replayed from DB; lagged live events
-                    // are a minor gap in the tail (tokens may be duplicated on
-                    // reconnect from DB, which is acceptable).
-                }
-            }
-        }
-    });
-
-    sse_response(sse_rx)
-}
-
-// -- Quick-message handler --------------------------------------------------
-
-/// `POST /api/quick-message` — create a new conversation, send a message, and
-/// return the complete assistant response as a synchronous JSON reply.
-///
-/// This endpoint is designed for machine clients (Apple Shortcuts, Siri App
-/// Intents, curl, webhooks) that need a single request/response round-trip.
-#[utoipa::path(
-    post,
-    path = "/api/quick-message",
-    operation_id = "create_quick_message",
-    tag = "conversations",
-    request_body = QuickMessageRequest,
-    responses(
-        (status = 201, description = "Message processed, conversation created", body = QuickMessageResponse),
-        (status = 400, description = "Empty message"),
-        (status = 401, description = "Unauthorized"),
-        (status = 500, description = "Internal server error"),
-    ),
-    security(("bearer_token" = []))
-)]
-pub async fn quick_message(
-    State(state): State<ApiState>,
-    Json(body): Json<QuickMessageRequest>,
-) -> Response {
-    let content = body.message.trim().to_string();
-    if content.is_empty() {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(serde_json::json!({"error": "Message cannot be empty"})),
-        )
-            .into_response();
-    }
-
-    // Resolve persona: use explicit persona_id if valid, otherwise active.
-    let agent_id = if let Some(ref pid) = body.persona_id {
-        let persona_store = assistant_storage::personas::PersonaStore::new(state.pool.clone());
-        if persona_store.get(pid).await.ok().flatten().is_some() {
-            pid.clone()
-        } else {
-            state.agent_id.read().await.clone()
-        }
-    } else {
-        state.agent_id.read().await.clone()
-    };
-    let store = ConversationStore::for_agent(state.pool.clone(), &agent_id)
-        .with_broadcaster(state.conversation_broadcaster.clone());
-
-    // Build the prompt: prepend context if provided.
-    let prompt = match body.context.as_deref() {
-        Some(ctx) if !ctx.trim().is_empty() => {
-            format!("<context>\n{}\n</context>\n\n{}", ctx.trim(), content)
-        }
-        _ => content.clone(),
-    };
-
-    // Create a new conversation with no title; the title-generator worker
-    // (consumer of `turn.result`) will produce one once the conversation has
-    // accumulated enough material.
-    let conv = match store.create_conversation(None).await {
-        Ok(c) => c,
-        Err(e) => {
-            warn!("quick-message: failed to create conversation: {e}");
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(serde_json::json!({"error": "Failed to create conversation"})),
-            )
-                .into_response();
-        }
-    };
-
-    // Submit the turn and await the full result (synchronous).
-    let result = state
-        .orchestrator
-        .submit_turn(&prompt, conv.id, Interface::Web, None)
-        .await;
-
-    match result {
-        Ok(turn_result) => (
-            StatusCode::CREATED,
-            Json(QuickMessageResponse {
-                conversation_id: conv.id,
-                message_id: turn_result.message_id,
-                answer: turn_result.answer,
-            }),
-        )
-            .into_response(),
-        Err(e) => {
-            warn!("quick-message: orchestrator error for {}: {e}", conv.id);
-            // Clean up the orphan conversation.
-            if let Err(del_err) = store.delete_conversation(conv.id).await {
-                warn!(
-                    "quick-message: failed to delete orphan conversation {}: {del_err}",
-                    conv.id
-                );
-            }
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(serde_json::json!({"error": "Failed to process message"})),
-            )
-                .into_response()
-        }
-    }
-}
-
-// -- Voice handler ----------------------------------------------------------
-
-/// Multipart form body for `POST /api/conversations/{id}/voice`.
-#[derive(utoipa::ToSchema)]
-#[allow(dead_code)]
-struct VoiceUploadForm {
-    /// Raw audio bytes (opus/aac/webm/wav …).
-    #[schema(format = Binary, content_encoding = "binary")]
-    audio: Vec<u8>,
-}
-
-/// `POST /api/conversations/{id}/voice` — upload audio, transcribe it, run
-/// through the orchestrator, and stream the response as SSE.
-#[utoipa::path(
-    post,
-    path = "/api/conversations/{id}/voice",
-    tag = "conversations",
-    params(("id" = Uuid, Path, description = "Conversation ID")),
-    request_body(
-        content_type = "multipart/form-data",
-        content = inline(VoiceUploadForm),
-    ),
-    responses(
-        (status = 200, description = "SSE stream of assistant response", content_type = "text/event-stream"),
-        (status = 400, description = "Bad request"),
-        (status = 401, description = "Unauthorized"),
-        (status = 503, description = "Voice not configured"),
-    ),
-    security(("bearer_token" = []))
-)]
-pub async fn send_voice_message(
-    State(state): State<ApiState>,
-    Path(id): Path<String>,
-    mut multipart: Multipart,
-) -> Response {
-    let conv_id = match Uuid::parse_str(&id) {
-        Ok(u) => u,
-        Err(_) => return (StatusCode::BAD_REQUEST, "Invalid conversation ID").into_response(),
-    };
-
-    let transcription_provider = match &state.transcription_provider {
-        Some(p) => p.clone(),
-        None => {
-            return (
-                StatusCode::SERVICE_UNAVAILABLE,
-                "Voice transcription not configured",
-            )
-                .into_response();
-        }
-    };
-
-    // Parse the multipart body — expect a single "audio" field.
-    let mut audio_bytes: Option<Vec<u8>> = None;
-    let mut mime_type = "audio/webm".to_string();
-
-    while let Ok(Some(field)) = multipart.next_field().await {
-        if field.name() == Some("audio") {
-            if let Some(ct) = field.content_type() {
-                mime_type = ct.to_string();
-            }
-            match field.bytes().await {
-                Ok(b) => {
-                    // Enforce 25 MB limit.
-                    if b.len() > 25 * 1024 * 1024 {
-                        return (StatusCode::BAD_REQUEST, "Audio too large (max 25 MB)")
-                            .into_response();
-                    }
-                    audio_bytes = Some(b.to_vec());
-                }
-                Err(e) => {
-                    warn!("Failed to read audio field: {e}");
-                    return (StatusCode::BAD_REQUEST, "Failed to read audio data").into_response();
-                }
-            }
-        }
-    }
-
-    let audio_bytes = match audio_bytes {
-        Some(b) if !b.is_empty() => b,
-        _ => return (StatusCode::BAD_REQUEST, "Missing audio field").into_response(),
-    };
-
-    // Validate MIME type — must be audio/*.
-    if !mime_type.starts_with("audio/") {
-        return (
-            StatusCode::BAD_REQUEST,
-            "Invalid MIME type — expected audio/*",
-        )
-            .into_response();
-    }
-
-    // Verify the conversation exists.
-    let agent_id = state.agent_id.read().await.clone();
-    let store = ConversationStore::for_agent(state.pool.clone(), &agent_id);
-    match store.get_conversation(conv_id).await {
-        Ok(None) => return (StatusCode::NOT_FOUND, "Conversation not found").into_response(),
-        Err(e) => {
-            warn!("Failed to check conversation {conv_id}: {e}");
-            return (StatusCode::INTERNAL_SERVER_ERROR, "Database error").into_response();
-        }
-        _ => {}
-    }
-
-    // Transcribe.
-    let transcript = match transcription_provider
-        .transcribe(TranscriptionRequest {
-            audio_data: audio_bytes,
-            mime_type: mime_type.clone(),
-            filename: None,
-            language: None,
-        })
-        .await
-    {
-        Ok(r) => r.text,
-        Err(e) => {
-            warn!("Transcription failed: {e}");
-            return (StatusCode::INTERNAL_SERVER_ERROR, "Transcription failed").into_response();
-        }
-    };
-
-    if transcript.trim().is_empty() {
-        return (StatusCode::BAD_REQUEST, "Transcription returned empty text").into_response();
-    }
-
-    // Run through orchestrator (same flow as send_message).
-    let (sse_tx, sse_rx) = mpsc::channel::<Result<Event, Infallible>>(64);
-    let (event_tx, mut event_rx) = mpsc::channel::<OrchestratorEvent>(64);
-
-    state
-        .orchestrator
-        .register_token_sink(conv_id, event_tx)
-        .await;
-
-    let orchestrator = state.orchestrator.clone();
-    let content = transcript.clone();
-    let turn_result_rx = {
-        let (tx, rx) =
-            tokio::sync::oneshot::channel::<anyhow::Result<assistant_runtime::TurnResult>>();
-        tokio::spawn(async move {
-            let result = orchestrator
-                .submit_turn(&content, conv_id, Interface::Web, None)
-                .await;
-            let _ = tx.send(result);
-        });
-        rx
-    };
-
-    // Emit the user transcript as the first SSE event so the UI can display it.
-    let transcript_event = serde_json::json!({ "role": "user", "content": transcript });
-    let _ = sse_tx
-        .send(Ok(Event::default()
-            .event("transcript")
-            .data(transcript_event.to_string())))
-        .await;
-
-    let push_dispatcher_for_sse = state.push_dispatcher.clone();
-    let conv_id_for_push = conv_id;
-    tokio::spawn(async move {
-        let mut full_text = String::new();
-
-        while let Some(orch_event) = event_rx.recv().await {
-            let sse_event = match orch_event {
-                OrchestratorEvent::Token(token) => {
-                    full_text.push_str(&token);
-                    Event::default().event("token").data(token)
-                }
-                OrchestratorEvent::Status {
-                    message,
-                    tool_call_id,
-                } => {
-                    let mut data = serde_json::json!({"message": message});
-                    if let Some(id) = tool_call_id {
-                        data["tool_call_id"] = serde_json::Value::String(id);
-                    }
-                    Event::default().event("status").data(data.to_string())
-                }
-                OrchestratorEvent::ToolResult {
-                    tool_name,
-                    status,
-                    arguments,
-                    result,
-                    tool_call_id,
-                } => {
-                    let mut data = serde_json::json!({
-                        "tool_name": tool_name,
-                        "status": status,
-                    });
-                    if let Some(args) = arguments {
-                        data["arguments"] = args;
-                    }
-                    if let Some(res) = result {
-                        data["result"] = serde_json::Value::String(res);
-                    }
-                    if let Some(id) = tool_call_id {
-                        data["tool_call_id"] = serde_json::Value::String(id);
-                    }
-                    Event::default().event("tool_result").data(data.to_string())
-                }
-                OrchestratorEvent::SkillComplete {
-                    skill_name,
-                    success,
-                    summary,
-                } => {
-                    let data = serde_json::json!({
-                        "skill_name": skill_name,
-                        "success": success,
-                        "summary": summary,
-                    });
-                    Event::default()
-                        .event("skill_complete")
-                        .data(data.to_string())
-                }
-                OrchestratorEvent::AgentError { message } => {
-                    Event::default().event("agent_error").data(message)
-                }
-                OrchestratorEvent::Thinking(content) => {
-                    let data = serde_json::json!({"content": content});
-                    Event::default().event("thinking").data(data.to_string())
-                }
-                OrchestratorEvent::SubagentStarted { agent_id, task } => {
-                    let data = serde_json::json!({"agent_id": agent_id, "task": task});
-                    Event::default()
-                        .event("subagent_started")
-                        .data(data.to_string())
-                }
-                OrchestratorEvent::SubagentCompleted {
-                    agent_id,
-                    status,
-                    summary,
-                } => {
-                    let data = serde_json::json!({"agent_id": agent_id, "status": status, "summary": summary});
-                    Event::default()
-                        .event("subagent_completed")
-                        .data(data.to_string())
-                }
-                OrchestratorEvent::SubagentEvent { agent_id, inner } => {
-                    let (inner_type, inner_data) = match inner.as_ref() {
-                        OrchestratorEvent::Token(t) => ("token", serde_json::json!({"content": t})),
-                        OrchestratorEvent::Thinking(t) => {
-                            ("thinking", serde_json::json!({"content": t}))
-                        }
-                        OrchestratorEvent::Status {
-                            message,
-                            tool_call_id,
-                        } => {
-                            let mut d = serde_json::json!({"message": message});
-                            if let Some(id) = tool_call_id {
-                                d["tool_call_id"] = serde_json::Value::String(id.clone());
-                            }
-                            ("status", d)
-                        }
-                        OrchestratorEvent::ToolResult {
-                            tool_name,
-                            status,
-                            arguments,
-                            result,
-                            tool_call_id,
-                        } => {
-                            let mut d = serde_json::json!({"tool_name": tool_name, "status": status, "arguments": arguments, "result": result});
-                            if let Some(id) = tool_call_id {
-                                d["tool_call_id"] = serde_json::Value::String(id.clone());
-                            }
-                            ("tool_result", d)
-                        }
-                        other => ("event", serde_json::json!({"type": format!("{other:?}")})),
-                    };
-                    let event_type = format!("subagent_{inner_type}");
-                    let data = serde_json::json!({"agent_id": agent_id, "event_type": inner_type, "data": inner_data});
-                    Event::default().event(&event_type).data(data.to_string())
-                }
-                OrchestratorEvent::AudioReady { audio_id } => {
-                    let data = serde_json::json!({
-                        "audio_id": audio_id,
-                        "auto_play": true,
-                    });
-                    Event::default().event("audio_ready").data(data.to_string())
-                }
-            };
-            if sse_tx.send(Ok(sse_event)).await.is_err() {
-                return;
-            }
-        }
-
-        let (reply_text, reply_message_id, reply_attachment_ids) = match turn_result_rx.await {
-            Ok(Ok(result)) => (result.answer, result.message_id, result.attachment_ids),
-            Ok(Err(_)) | Err(_) => (full_text, None, vec![]),
-        };
-
-        let mut done_data = serde_json::json!({
-            "role": "assistant",
-            "content": reply_text,
-        });
-        if let Some(mid) = reply_message_id {
-            done_data["message_id"] = serde_json::Value::String(mid.to_string());
-        }
-        if !reply_attachment_ids.is_empty() {
-            done_data["attachment_ids"] = serde_json::json!(
-                reply_attachment_ids
-                    .iter()
-                    .map(|id| id.to_string())
-                    .collect::<Vec<_>>()
-            );
-        }
-        let done = Event::default().event("done").data(done_data.to_string());
-        let _ = sse_tx.send(Ok(done)).await;
-
-        if let Some(dispatcher) = push_dispatcher_for_sse {
-            let body = reply_text.chars().take(80).collect::<String>();
-            let conv_id_str = conv_id_for_push.to_string();
-            if let Err(e) = dispatcher
-                .send_to_all("New message", &body, Some(&conv_id_str))
-                .await
-            {
-                warn!("Push dispatch failed: {e}");
-            }
-        }
-    });
-
-    sse_response(sse_rx)
-}
-
-// -- Tests -------------------------------------------------------------------
-
-#[cfg(test)]
-mod tests {
-    use std::sync::Arc;
-
-    use axum::body::Body;
-    use axum::http::{Request, StatusCode};
-    use chrono::Utc;
-    use tower::ServiceExt;
-    use uuid::Uuid;
-    use wiremock::MockServer;
-
-    use assistant_storage::{ConversationStore, SkillRegistry, StorageLayer};
-
-    use super::test_helpers::*;
-    use super::*;
-
-    // -- POST /conversations/{id}/messages — error paths ----------------------
-
-    #[tokio::test]
-    async fn send_message_bad_uuid_returns_400() {
-        let server = MockServer::start().await;
-        let (state, _) = test_state(&server.uri()).await;
-
-        let resp = app(state)
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/conversations/not-a-uuid/messages")
-                    .header("Content-Type", "application/json")
-                    .body(Body::from(r#"{"message":"hi"}"#))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
-    }
-
-    #[tokio::test]
-    async fn send_message_empty_body_returns_400() {
-        let server = MockServer::start().await;
-        let (state, storage) = test_state(&server.uri()).await;
-
-        let store = ConversationStore::for_agent(storage.pool.clone(), "default");
-        let conv = store.create_conversation(Some("T")).await.unwrap();
-        let id = conv.id;
-
-        let resp = app(state)
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri(format!("/conversations/{id}/messages"))
-                    .header("Content-Type", "application/json")
-                    .body(Body::from(r#"{"message":"   "}"#))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
-    }
-
-    #[tokio::test]
-    async fn test_send_message_first_turn_leaves_title_null() {
-        let server = MockServer::start().await;
-        mount_llm_reply(&server, "Hello world").await;
-
-        let (state, storage) = test_state(&server.uri()).await;
-        let store = ConversationStore::for_agent(storage.pool.clone(), "default");
-        let conv = store.create_conversation(None).await.unwrap();
-        assert!(conv.title.is_none(), "precondition: conversation untitled");
-        assert!(!conv.title_locked, "precondition: conversation unlocked");
-        let id = conv.id;
-
-        let resp = app(state)
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri(format!("/conversations/{id}/messages"))
-                    .header("Content-Type", "application/json")
-                    .body(Body::from(
-                        r#"{"message":"What should we have for dinner tonight?"}"#,
-                    ))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(resp.status(), StatusCode::OK);
-        // Drain the SSE stream so the turn completes before we re-read.
-        let _ = body_bytes(resp.into_body()).await;
-
-        let loaded = store.get_conversation(id).await.unwrap().unwrap();
-        assert!(
-            loaded.title.is_none(),
-            "title must remain NULL after first message — got {:?}",
-            loaded.title
-        );
-        assert!(
-            !loaded.title_locked,
-            "title_locked must remain false; the worker (not send_message) is responsible for titling"
-        );
-    }
-
-    #[tokio::test]
-    async fn send_message_unknown_conversation_returns_404() {
-        let server = MockServer::start().await;
-        let (state, _) = test_state(&server.uri()).await;
-        let id = uuid::Uuid::new_v4();
-
-        let resp = app(state)
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri(format!("/conversations/{id}/messages"))
-                    .header("Content-Type", "application/json")
-                    .body(Body::from(r#"{"message":"hello"}"#))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(resp.status(), StatusCode::NOT_FOUND);
-    }
-
-    #[tokio::test]
-    async fn send_message_streams_sse_tokens() {
-        let server = MockServer::start().await;
-        mount_llm_reply(&server, "Hello world").await;
-
-        let (state, storage) = test_state(&server.uri()).await;
-        let store = ConversationStore::for_agent(storage.pool.clone(), "default");
-        let conv = store.create_conversation(Some("Stream")).await.unwrap();
-        let id = conv.id;
-
-        let resp = app(state)
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri(format!("/conversations/{id}/messages"))
-                    .header("Content-Type", "application/json")
-                    .body(Body::from(r#"{"message":"ping"}"#))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(resp.status(), StatusCode::OK);
-        assert_eq!(
-            resp.headers()
-                .get("content-type")
-                .and_then(|v| v.to_str().ok()),
-            Some("text/event-stream")
-        );
-        // Consume enough of the body to confirm SSE framing is present.
-        let body = body_bytes(resp.into_body()).await;
-        let text = String::from_utf8_lossy(&body);
-        assert!(
-            text.contains("event:") || text.contains("data:"),
-            "expected SSE framing, got: {text}"
-        );
-    }
-
-    #[tokio::test]
-    async fn send_message_done_event_contains_message_id() {
-        let server = MockServer::start().await;
-        mount_llm_reply(&server, "Hello world").await;
-
-        let (state, storage) = test_state(&server.uri()).await;
-        let store = ConversationStore::for_agent(storage.pool.clone(), "default");
-        let conv = store.create_conversation(Some("MsgId")).await.unwrap();
-        let id = conv.id;
-
-        let resp = app(state)
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri(format!("/conversations/{id}/messages"))
-                    .header("Content-Type", "application/json")
-                    .body(Body::from(r#"{"message":"hi"}"#))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(resp.status(), StatusCode::OK);
-        let body = body_bytes(resp.into_body()).await;
-        let text = String::from_utf8_lossy(&body);
-        // Find the done event data line (role == "assistant").
-        let done_line = text
-            .lines()
-            .find(|l| l.starts_with("data:") && l.contains("\"assistant\""))
-            .expect("expected a done data line with assistant role");
-        let data = done_line.trim_start_matches("data:").trim();
-        let json: serde_json::Value = serde_json::from_str(data).expect("valid JSON");
-        assert_eq!(json["role"], "assistant", "done role should be assistant");
-        assert!(
-            json.get("message_id").and_then(|v| v.as_str()).is_some(),
-            "done event should contain message_id, got: {json}"
-        );
-        // message_id must be a valid UUID.
-        let mid = json["message_id"].as_str().unwrap();
-        assert!(
-            uuid::Uuid::parse_str(mid).is_ok(),
-            "message_id should be a valid UUID, got: {mid}"
-        );
-    }
-
-    #[tokio::test]
-    async fn send_voice_message_done_event_contains_message_id() {
-        let server = MockServer::start().await;
-        mount_llm_reply(&server, "Voice reply").await;
-
-        let (state, storage) = test_state(&server.uri()).await;
-        let store = ConversationStore::for_agent(storage.pool.clone(), "default");
-        let conv = store.create_conversation(Some("VoiceMsgId")).await.unwrap();
-
-        // Wire up the stub transcription provider.
-        let state = ApiState {
-            transcription_provider: Some(Arc::new(StubTranscriptionProvider {
-                transcript: "hello from voice".to_string(),
-            })),
-            ..state
-        };
-
-        let boundary = "testboundary";
-        let body_str = format!(
-            "--{boundary}\r\nContent-Disposition: form-data; name=\"audio\"; filename=\"audio.webm\"\r\nContent-Type: audio/webm\r\n\r\nfakeaudio\r\n--{boundary}--\r\n",
-        );
-
-        let resp = app(state)
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri(format!("/conversations/{}/voice", conv.id))
-                    .header(
-                        "content-type",
-                        format!("multipart/form-data; boundary={boundary}"),
-                    )
-                    .body(Body::from(body_str))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(resp.status(), StatusCode::OK);
-        let body = body_bytes(resp.into_body()).await;
-        let text = String::from_utf8_lossy(&body);
-
-        // The stream should contain a transcript event with the stub text.
-        assert!(
-            text.contains("transcript"),
-            "expected transcript event in SSE stream, got: {text}"
-        );
-
-        // Find the done event data line (role == "assistant").
-        let done_line = text
-            .lines()
-            .find(|l| l.starts_with("data:") && l.contains("\"assistant\""))
-            .expect("expected a done data line with assistant role");
-        let data = done_line.trim_start_matches("data:").trim();
-        let json: serde_json::Value = serde_json::from_str(data).expect("valid JSON");
-        assert_eq!(json["role"], "assistant");
-        assert!(
-            json.get("message_id").and_then(|v| v.as_str()).is_some(),
-            "voice done event should contain message_id, got: {json}"
-        );
-        let mid = json["message_id"].as_str().unwrap();
-        assert!(
-            uuid::Uuid::parse_str(mid).is_ok(),
-            "message_id should be a valid UUID, got: {mid}"
-        );
-    }
-
-    // -- POST /conversations/{id}/voice — MIME validation & size limit ---------
-
-    #[tokio::test]
-    async fn voice_upload_without_transcription_provider_returns_503() {
-        let server = MockServer::start().await;
-        let (state, storage) = test_state(&server.uri()).await;
-        let store = ConversationStore::for_agent(storage.pool.clone(), "default");
-        let conv = store.create_conversation(Some("test")).await.unwrap();
-
-        // Build a minimal multipart body
-        let boundary = "testboundary";
-        let body_bytes = format!(
-            "--{boundary}\r\nContent-Disposition: form-data; name=\"audio\"; filename=\"audio.webm\"\r\nContent-Type: audio/webm\r\n\r\nfakeaudio\r\n--{boundary}--\r\n",
-            boundary = boundary
-        );
-
-        let resp = app(state)
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri(format!("/conversations/{}/voice", conv.id))
-                    .header(
-                        "content-type",
-                        format!("multipart/form-data; boundary={boundary}"),
-                    )
-                    .body(Body::from(body_bytes))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(resp.status(), StatusCode::SERVICE_UNAVAILABLE);
-    }
-
-    #[tokio::test]
-    async fn voice_upload_invalid_conversation_id_returns_400() {
-        let server = MockServer::start().await;
-        let (state, _) = test_state(&server.uri()).await;
-
-        let boundary = "testboundary";
-        let body_bytes = format!(
-            "--{boundary}\r\nContent-Disposition: form-data; name=\"audio\"\r\nContent-Type: audio/webm\r\n\r\nfake\r\n--{boundary}--\r\n",
-            boundary = boundary
-        );
-
-        let resp = app(state)
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/conversations/not-a-uuid/voice")
-                    .header(
-                        "content-type",
-                        format!("multipart/form-data; boundary={boundary}"),
-                    )
-                    .body(Body::from(body_bytes))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
-    }
-
-    // -- GET /audio/{id} -------------------------------------------------------
-
-    // -- stream_run_events tests -----------------------------------------------
-
-    /// Seed the event store directly and replay via the endpoint.
-    #[tokio::test]
-    async fn stream_run_events_replays_completed_run() {
-        let (state, _storage) = event_log_state().await;
-        let conv_id = Uuid::new_v4().to_string();
-        let run_id = Uuid::new_v4().to_string();
-
-        // Seed events into the store.
-        state
-            .event_store
-            .append_event(
-                &run_id,
-                &conv_id,
-                0,
-                "run_started",
-                &serde_json::json!({"run_id": run_id}),
-            )
-            .await
-            .unwrap();
-        state
-            .event_store
-            .append_event(
-                &run_id,
-                &conv_id,
-                1,
-                "token",
-                &serde_json::json!({"token": "hello"}),
-            )
-            .await
-            .unwrap();
-        state
-            .event_store
-            .append_event(
-                &run_id,
-                &conv_id,
-                2,
-                "done",
-                &serde_json::json!({"content": "hello"}),
-            )
-            .await
-            .unwrap();
-
-        let app = app(state);
-        let req = Request::builder()
-            .uri(format!(
-                "/conversations/{conv_id}/runs/{run_id}/events/stream"
-            ))
-            .header("Authorization", "Bearer test-token")
-            .body(Body::empty())
-            .unwrap();
-        let resp = app.oneshot(req).await.unwrap();
-        assert_eq!(resp.status(), StatusCode::OK);
-        assert_eq!(
-            resp.headers()
-                .get("content-type")
-                .and_then(|v| v.to_str().ok()),
-            Some("text/event-stream")
-        );
-        let body = body_bytes(resp.into_body()).await;
-        let text = String::from_utf8_lossy(&body);
-        assert!(
-            text.contains("run_started"),
-            "should include run_started: {text}"
-        );
-        assert!(text.contains("token"), "should include token event: {text}");
-        assert!(text.contains("done"), "should include done event: {text}");
-    }
-
-    #[tokio::test]
-    async fn stream_run_events_returns_404_for_unknown_run() {
-        let (state, _storage) = event_log_state().await;
-        let conv_id = Uuid::new_v4().to_string();
-        let run_id = Uuid::new_v4().to_string(); // never seeded
-
-        let app = app(state);
-        let req = Request::builder()
-            .uri(format!(
-                "/conversations/{conv_id}/runs/{run_id}/events/stream"
-            ))
-            .header("Authorization", "Bearer test-token")
-            .body(Body::empty())
-            .unwrap();
-        let resp = app.oneshot(req).await.unwrap();
-        assert_eq!(resp.status(), StatusCode::NOT_FOUND);
-    }
-
-    #[tokio::test]
-    async fn stream_run_events_since_skips_earlier_events() {
-        let (state, _storage) = event_log_state().await;
-        let conv_id = Uuid::new_v4().to_string();
-        let run_id = Uuid::new_v4().to_string();
-
-        for i in 0i64..5 {
-            state
-                .event_store
-                .append_event(
-                    &run_id,
-                    &conv_id,
-                    i,
-                    if i == 4 { "done" } else { "token" },
-                    &serde_json::json!({"token": format!("t{i}")}),
-                )
-                .await
-                .unwrap();
-        }
-
-        let app = app(state);
-        let req = Request::builder()
-            .uri(format!(
-                "/conversations/{conv_id}/runs/{run_id}/events/stream?since=3"
-            ))
-            .header("Authorization", "Bearer test-token")
-            .body(Body::empty())
-            .unwrap();
-        let resp = app.oneshot(req).await.unwrap();
-        assert_eq!(resp.status(), StatusCode::OK);
-        let body = body_bytes(resp.into_body()).await;
-        let text = String::from_utf8_lossy(&body);
-        // Only events 3 and 4 should appear.
-        assert!(text.contains("t3"), "should contain seq 3: {text}");
-        assert!(text.contains("done"), "should contain done: {text}");
-        assert!(!text.contains("t0"), "seq 0 should be skipped: {text}");
-    }
-
-    /// Client connects while the run is still active in the broadcaster.
-    /// It should receive the replayed DB event first, then live events as they
-    /// arrive, and close when "done" is broadcast.
-    #[tokio::test]
-    async fn stream_run_events_tails_live_broadcaster() {
-        let (state, _storage) = event_log_state().await;
-        let conv_id = Uuid::new_v4().to_string();
-        let run_id = Uuid::new_v4();
-        let run_id_str = run_id.to_string();
-
-        // Seed one already-persisted event (the run_started from before client connects).
-        state
-            .event_store
-            .append_event(
-                &run_id_str,
-                &conv_id,
-                0,
-                "run_started",
-                &serde_json::json!({"run_id": run_id_str}),
-            )
-            .await
-            .unwrap();
-
-        // Register the run as active in the broadcaster.
-        let broadcast_tx = state.run_broadcaster.start_run(run_id).await;
-        // Clone the broadcaster before state is moved into `app()`.
-        let broadcaster = state.run_broadcaster.clone();
-
-        // Spawn a task that delivers live events after a brief delay so the
-        // HTTP handler has time to subscribe before events are sent.
-        let run_id_for_task = run_id;
-        tokio::spawn(async move {
-            tokio::time::sleep(tokio::time::Duration::from_millis(25)).await;
-            let _ = broadcast_tx.send(assistant_storage::LiveEvent {
-                sequence: 1,
-                event_type: "token".to_string(),
-                payload: serde_json::json!({"token": "live-word"}),
-            });
-            let _ = broadcast_tx.send(assistant_storage::LiveEvent {
-                sequence: 2,
-                event_type: "done".to_string(),
-                payload: serde_json::json!({"content": "live-word", "role": "assistant"}),
-            });
-            // Drop sender and remove from registry so the stream closes.
-            broadcaster.finish_run(&run_id_for_task).await;
-        });
-
-        let app = app(state);
-        let req = Request::builder()
-            .uri(format!(
-                "/conversations/{conv_id}/runs/{run_id_str}/events/stream"
-            ))
-            .header("Authorization", "Bearer test-token")
-            .body(Body::empty())
-            .unwrap();
-        let resp = app.oneshot(req).await.unwrap();
-        assert_eq!(resp.status(), StatusCode::OK);
-        assert_eq!(
-            resp.headers()
-                .get("content-type")
-                .and_then(|v| v.to_str().ok()),
-            Some("text/event-stream"),
-        );
-        let body = body_bytes(resp.into_body()).await;
-        let text = String::from_utf8_lossy(&body);
-        assert!(
-            text.contains("run_started"),
-            "replayed DB event missing: {text}"
-        );
-        assert!(
-            text.contains("live-word"),
-            "live token from broadcaster missing: {text}"
-        );
-        assert!(text.contains("done"), "done event missing: {text}");
-    }
-
-    /// Race-condition path: broadcaster is active but no DB events have been
-    /// persisted yet (run_started event is still in-flight). The handler should
-    /// recognise the run as active and switch to live-tail mode.
-    #[tokio::test]
-    async fn stream_run_events_live_only_when_no_db_events_yet() {
-        let (state, _storage) = event_log_state().await;
-        let conv_id = Uuid::new_v4().to_string();
-        let run_id = Uuid::new_v4();
-        let run_id_str = run_id.to_string();
-
-        // Register as active — no DB events at all.
-        let broadcast_tx = state.run_broadcaster.start_run(run_id).await;
-        let broadcaster = state.run_broadcaster.clone();
-
-        let run_id_for_task = run_id;
-        tokio::spawn(async move {
-            tokio::time::sleep(tokio::time::Duration::from_millis(25)).await;
-            let _ = broadcast_tx.send(assistant_storage::LiveEvent {
-                sequence: 0,
-                event_type: "token".to_string(),
-                payload: serde_json::json!({"token": "first-token"}),
-            });
-            let _ = broadcast_tx.send(assistant_storage::LiveEvent {
-                sequence: 1,
-                event_type: "done".to_string(),
-                payload: serde_json::json!({"content": "first-token", "role": "assistant"}),
-            });
-            broadcaster.finish_run(&run_id_for_task).await;
-        });
-
-        let app = app(state);
-        let req = Request::builder()
-            .uri(format!(
-                "/conversations/{conv_id}/runs/{run_id_str}/events/stream"
-            ))
-            .header("Authorization", "Bearer test-token")
-            .body(Body::empty())
-            .unwrap();
-        let resp = app.oneshot(req).await.unwrap();
-        assert_eq!(resp.status(), StatusCode::OK);
-        let body = body_bytes(resp.into_body()).await;
-        let text = String::from_utf8_lossy(&body);
-        assert!(
-            text.contains("first-token"),
-            "live token missing for early-connect case: {text}"
-        );
-        assert!(text.contains("done"), "done event missing: {text}");
-    }
-
-    /// Verify that the send_message response carries the X-Run-Id header so
-    /// clients can capture the run identifier even if the SSE stream is dropped
-    /// before the run_started event arrives.
-    #[tokio::test]
-    async fn send_message_sets_x_run_id_header() {
-        let server = MockServer::start().await;
-        mount_llm_reply(&server, "Test reply").await;
-
-        let (state, storage) = test_state(&server.uri()).await;
-        let store = ConversationStore::for_agent(storage.pool.clone(), "default");
-        let conv = store
-            .create_conversation(Some("RunId header"))
-            .await
-            .unwrap();
-        let id = conv.id;
-
-        let resp = app(state)
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri(format!("/conversations/{id}/messages"))
-                    .header("Content-Type", "application/json")
-                    .body(Body::from(r#"{"message":"hello"}"#))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(resp.status(), StatusCode::OK);
-        let run_id_header = resp
-            .headers()
-            .get("x-run-id")
-            .and_then(|v| v.to_str().ok())
-            .expect("X-Run-Id header must be present");
-        assert!(
-            Uuid::parse_str(run_id_header).is_ok(),
-            "X-Run-Id should be a valid UUID, got: {run_id_header}"
-        );
-    }
-
-    /// After events are pruned (TTL), the endpoint returns 404.
-    /// (We do not distinguish 404 vs 410 in the current implementation —
-    /// both expired and never-existed runs return 404.)
-    #[tokio::test]
-    async fn stream_run_events_returns_404_for_pruned_run() {
-        let sl = Arc::new(StorageLayer::new_in_memory().await.unwrap());
-        // Use a very short TTL (already negative = instantly expired).
-        let short_ttl_store = assistant_storage::ConversationEventStore::with_ttl(
-            sl.pool.clone(),
-            chrono::Duration::seconds(-1),
-        );
-
-        let run_id = Uuid::new_v4();
-        let run_id_str = run_id.to_string();
-        let conv_id = Uuid::new_v4().to_string();
-
-        // Append an event; it is already expired on insert.
-        short_ttl_store
-            .append_event(
-                &run_id_str,
-                &conv_id,
-                0,
-                "run_started",
-                &serde_json::json!({"run_id": run_id_str}),
-            )
-            .await
-            .unwrap();
-
-        // Prune — removes the expired row.
-        short_ttl_store.prune_expired().await.unwrap();
-
-        // Build a state that uses the same pool (events are now gone).
-        use assistant_core::types::agent::AssistantConfig;
-        use assistant_llm_provider::ollama::client::{LlmClient, LlmClientConfig};
-        use assistant_llm_provider::retry::RetryConfig;
-        use assistant_runtime::Orchestrator;
-        use assistant_tool_executor::ToolExecutor;
-        let mut config = AssistantConfig::default();
-        config.memory.enabled = false;
-        let registry = Arc::new(SkillRegistry::new(sl.pool.clone()).await.unwrap());
-        let llm = Arc::new(
-            LlmClient::new(LlmClientConfig {
-                model: "test".to_string(),
-                base_url: "http://127.0.0.1:1".to_string(),
-                timeout_secs: 1,
-                retry_config: RetryConfig::disabled(),
-            })
-            .unwrap(),
-        );
-        let executor = Arc::new(ToolExecutor::new(
-            sl.clone(),
-            llm.clone(),
-            registry.clone(),
-            Arc::new(config.clone()),
-        ));
-        let bus = Arc::new(sl.message_bus());
-        let orchestrator = Arc::new(Orchestrator::new(
-            llm,
-            sl.clone(),
-            executor,
-            registry,
-            bus,
-            &config,
-        ));
-        let orchestrator_ref = orchestrator.clone();
-        let default_model = orchestrator_ref.llm.model_name().to_string();
-        let state = ApiState {
-            pool: sl.pool.clone(),
-            agent_id: Arc::new(RwLock::new("default".to_string())),
-            orchestrator,
-            push_dispatcher: None,
-            transcription_provider: None,
-            tts_provider: None,
-            audio_store: Arc::new(crate::audio_store::AudioStore::new()),
-            event_store: short_ttl_store,
-            run_broadcaster: RunBroadcaster::new(),
-            attachment_store: AttachmentStore::new(sl.pool.clone()),
-            command_registry: Arc::new(CommandRegistry::new()),
-            command_event_store: CommandEventStore::new(sl.pool.clone()),
-            conversation_broadcaster: Arc::new(InMemoryConversationBroadcaster::new()),
-            conversation_configs: Arc::new(RwLock::new(HashMap::new())),
-            orchestrator_ref,
-            active_turns: Arc::new(RwLock::new(HashMap::new())),
-            default_model,
-        };
-
-        let app = app(state);
-        let req = Request::builder()
-            .uri(format!(
-                "/conversations/{conv_id}/runs/{run_id_str}/events/stream"
-            ))
-            .header("Authorization", "Bearer test-token")
-            .body(Body::empty())
-            .unwrap();
-        let resp = app.oneshot(req).await.unwrap();
-        // The run existed but events were pruned → currently 404.
-        assert_eq!(resp.status(), StatusCode::NOT_FOUND);
-    }
-
-    // -- POST /quick-message ---------------------------------------------------
-
-    #[tokio::test]
-    async fn quick_message_returns_201_with_answer() {
-        let server = MockServer::start().await;
-        mount_llm_reply(&server, "Hello from quick message").await;
-
-        let (state, _storage) = test_state(&server.uri()).await;
-
-        let resp = app(state)
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/quick-message")
-                    .header("Content-Type", "application/json")
-                    .body(Body::from(r#"{"message":"ping"}"#))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(resp.status(), StatusCode::CREATED);
-        let json = body_json(resp.into_body()).await;
-        assert!(
-            json["conversation_id"].as_str().is_some(),
-            "should have a conversation_id"
-        );
-        assert_eq!(
-            json["answer"].as_str().unwrap(),
-            "Hello from quick message",
-            "answer should match the LLM reply"
-        );
-    }
-
-    #[tokio::test]
-    async fn quick_message_empty_body_returns_400() {
-        let server = MockServer::start().await;
-        let (state, _) = test_state(&server.uri()).await;
-
-        let resp = app(state)
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/quick-message")
-                    .header("Content-Type", "application/json")
-                    .body(Body::from(r#"{"message":"   "}"#))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
-        let json = body_json(resp.into_body()).await;
-        assert_eq!(json["error"], "Message cannot be empty");
-    }
-
-    #[tokio::test]
-    async fn test_quick_message_creates_conversation_with_null_title() {
-        let server = MockServer::start().await;
-        mount_llm_reply(&server, "response").await;
-
-        let (state, storage) = test_state(&server.uri()).await;
-
-        let resp = app(state)
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/quick-message")
-                    .header("Content-Type", "application/json")
-                    .body(Body::from(
-                        r#"{"message":"What should I cook for dinner tonight?"}"#,
-                    ))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(resp.status(), StatusCode::CREATED);
-        let json = body_json(resp.into_body()).await;
-        let conv_id: Uuid = json["conversation_id"].as_str().unwrap().parse().unwrap();
-
-        // The legacy auto-truncation is gone — the conversation starts NULL
-        // and the title-generator worker is solely responsible for titling.
-        let store = ConversationStore::for_agent(storage.pool.clone(), "default");
-        let conv = store.get_conversation(conv_id).await.unwrap().unwrap();
-        assert!(
-            conv.title.is_none(),
-            "quick-message must not auto-truncate the title; got {:?}",
-            conv.title
-        );
-        assert!(!conv.title_locked);
-    }
-
-    #[tokio::test]
-    async fn quick_message_with_persona_id_routes_to_persona() {
-        let server = MockServer::start().await;
-        mount_llm_reply(&server, "persona answer").await;
-
-        let (state, storage) = test_state(&server.uri()).await;
-
-        // Create a persona so it can be resolved.
-        let persona_store = assistant_storage::personas::PersonaStore::new(storage.pool.clone());
-        persona_store.create("reviewer", "Reviewer").await.unwrap();
-
-        let resp = app(state)
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/quick-message")
-                    .header("Content-Type", "application/json")
-                    .body(Body::from(r#"{"message":"hello","persona_id":"reviewer"}"#))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(resp.status(), StatusCode::CREATED);
-        let json = body_json(resp.into_body()).await;
-        let conv_id: Uuid = json["conversation_id"].as_str().unwrap().parse().unwrap();
-
-        // Conversation should be scoped to the "reviewer" persona.
-        let store = ConversationStore::for_agent(storage.pool.clone(), "reviewer");
-        let conv = store.get_conversation(conv_id).await.unwrap();
-        assert!(
-            conv.is_some(),
-            "conversation should exist under 'reviewer' persona"
-        );
-    }
-
-    #[tokio::test]
-    async fn quick_message_with_invalid_persona_falls_back() {
-        let server = MockServer::start().await;
-        mount_llm_reply(&server, "fallback answer").await;
-
-        let (state, _storage) = test_state(&server.uri()).await;
-
-        let resp = app(state)
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/quick-message")
-                    .header("Content-Type", "application/json")
-                    .body(Body::from(
-                        r#"{"message":"hello","persona_id":"nonexistent"}"#,
-                    ))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        // Should succeed using fallback persona, not error.
-        assert_eq!(resp.status(), StatusCode::CREATED);
-        let json = body_json(resp.into_body()).await;
-        assert_eq!(json["answer"].as_str().unwrap(), "fallback answer");
-    }
-
-    #[tokio::test]
-    async fn quick_message_with_context_prepends_to_prompt() {
-        let server = MockServer::start().await;
-        mount_llm_reply(&server, "context answer").await;
-
-        let (state, _storage) = test_state(&server.uri()).await;
-
-        let resp = app(state)
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/quick-message")
-                    .header("Content-Type", "application/json")
-                    .body(Body::from(
-                        r#"{"message":"summarize this","context":"Lorem ipsum dolor sit amet"}"#,
-                    ))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(resp.status(), StatusCode::CREATED);
-        let json = body_json(resp.into_body()).await;
-        assert_eq!(json["answer"].as_str().unwrap(), "context answer");
-    }
-
-    #[tokio::test]
-    async fn quick_message_backward_compatible() {
-        let server = MockServer::start().await;
-        mount_llm_reply(&server, "compat answer").await;
-
-        let (state, _storage) = test_state(&server.uri()).await;
-
-        // Send only `message` — no persona_id, no context.
-        let resp = app(state)
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/quick-message")
-                    .header("Content-Type", "application/json")
-                    .body(Body::from(r#"{"message":"hi"}"#))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(resp.status(), StatusCode::CREATED);
-        let json = body_json(resp.into_body()).await;
-        assert_eq!(json["answer"].as_str().unwrap(), "compat answer");
-    }
-
-    // -- stream_conversations tests --------------------------------------------
-
-    #[tokio::test]
-    async fn stream_conversations_returns_snapshot() {
-        let (state, storage) = event_log_state().await;
-
-        // Create a conversation so the snapshot is non-empty.
-        let store = ConversationStore::for_agent(storage.pool.clone(), "default");
-        store.create_conversation(Some("Hello")).await.unwrap();
-
-        let resp = app(state)
-            .oneshot(
-                Request::builder()
-                    .uri("/conversations/stream")
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(resp.status(), StatusCode::OK);
-
-        let body = body_bytes(resp.into_body()).await;
-        let text = String::from_utf8_lossy(&body);
-
-        // Should contain a snapshot event.
-        assert!(
-            text.contains("event: snapshot"),
-            "response should contain snapshot event, got: {text}"
-        );
-        assert!(
-            text.contains("Hello"),
-            "snapshot should contain the conversation title"
-        );
-    }
-
-    #[tokio::test]
-    async fn stream_conversations_forwards_upserted_delta() {
-        let (state, _storage) = event_log_state().await;
-        let broadcaster = state.conversation_broadcaster.clone();
-        let pool = state.pool.clone();
-
-        // Spawn the SSE stream reader in a task.
-        let app = app(state);
-        let handle = tokio::spawn(async move {
-            let resp = app
-                .oneshot(
-                    Request::builder()
-                        .uri("/conversations/stream")
-                        .body(Body::empty())
-                        .unwrap(),
-                )
-                .await
-                .unwrap();
-            assert_eq!(resp.status(), StatusCode::OK);
-            body_bytes(resp.into_body()).await
-        });
-
-        // Give the stream a moment to subscribe.
-        tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-
-        // Create a conversation through a broadcaster-wired store.
-        let store = ConversationStore::for_agent(pool, "default").with_broadcaster(broadcaster);
-        store.create_conversation(Some("Delta Test")).await.unwrap();
-
-        // Give the stream a moment to forward the event, then drop broadcaster
-        // to close the channel so the SSE task finishes.
-        tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-        drop(store);
-
-        let body = tokio::time::timeout(std::time::Duration::from_secs(5), handle)
-            .await
-            .expect("stream should finish within timeout")
-            .expect("stream task should not panic");
-        let text = String::from_utf8_lossy(&body);
-
-        assert!(
-            text.contains("event: upserted"),
-            "stream should contain upserted event, got: {text}"
-        );
-        assert!(
-            text.contains("Delta Test"),
-            "upserted event should contain the conversation title"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_update_title_broadcasts_upserted_with_new_title() {
-        // Closes the loop the title-generator worker depends on: when the
-        // worker calls update_title on an existing conversation, an Upserted
-        // event MUST be broadcast carrying the new title so any subscribed
-        // UI re-renders without a refresh.  The wire-level forward to SSE
-        // is covered by `stream_conversations_forwards_upserted_delta`.
-        let (state, _storage) = event_log_state().await;
-        let broadcaster = state.conversation_broadcaster.clone();
-        let pool = state.pool.clone();
-
-        let store =
-            ConversationStore::for_agent(pool, "default").with_broadcaster(broadcaster.clone());
-        let conv = store.create_conversation(None).await.unwrap();
-        let conv_id = conv.id;
-
-        // Subscribe AFTER create so we only capture the update.
-        let mut rx = broadcaster.subscribe();
-        store
-            .update_title(conv_id, "Generated By Worker")
-            .await
-            .unwrap();
-
-        let event = tokio::time::timeout(std::time::Duration::from_secs(2), rx.recv())
-            .await
-            .expect("broadcast must arrive within 2s")
-            .expect("recv must not error");
-
-        match event {
-            assistant_storage::ConversationEvent::Upserted(record) => {
-                assert_eq!(record.id, conv_id);
-                assert_eq!(record.title.as_deref(), Some("Generated By Worker"));
-                assert!(
-                    record.title_locked,
-                    "broadcast after update_title must carry title_locked = true",
-                );
-            }
-            other => panic!("expected Upserted, got {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn stream_conversations_filters_by_agent_id() {
-        use assistant_storage::ConversationEvent;
-
-        let (state, _storage) = event_log_state().await;
-        let broadcaster = state.conversation_broadcaster.clone();
-
-        // Stream filtered to agent "work".
-        let app = app(state);
-        let handle = tokio::spawn(async move {
-            let resp = app
-                .oneshot(
-                    Request::builder()
-                        .uri("/conversations/stream?agent_id=work")
-                        .body(Body::empty())
-                        .unwrap(),
-                )
-                .await
-                .unwrap();
-            body_bytes(resp.into_body()).await
-        });
-
-        tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-
-        // Emit an event for agent "default" — should be filtered out.
-        broadcaster.emit(ConversationEvent::Upserted(
-            assistant_storage::ConversationRecord {
-                id: Uuid::new_v4(),
-                agent_id: "default".to_string(),
-                title: Some("DefaultConv".to_string()),
-                title_locked: true,
-                user_id: None,
-                created_at: Utc::now(),
-                updated_at: Utc::now(),
-            },
-        ));
-
-        // Emit an event for agent "work" — should pass through.
-        broadcaster.emit(ConversationEvent::Upserted(
-            assistant_storage::ConversationRecord {
-                id: Uuid::new_v4(),
-                agent_id: "work".to_string(),
-                title: Some("WorkConv".to_string()),
-                title_locked: true,
-                user_id: None,
-                created_at: Utc::now(),
-                updated_at: Utc::now(),
-            },
-        ));
-
-        tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-        drop(broadcaster);
-
-        let body = tokio::time::timeout(std::time::Duration::from_secs(5), handle)
-            .await
-            .expect("timeout")
-            .expect("no panic");
-        let text = String::from_utf8_lossy(&body);
-
-        assert!(
-            !text.contains("DefaultConv"),
-            "should NOT contain event for wrong agent, got: {text}"
-        );
-        assert!(
-            text.contains("WorkConv"),
-            "should contain event for filtered agent, got: {text}"
-        );
-    }
-
-    #[tokio::test]
-    async fn stream_conversations_e2e_create_and_delete() {
-        let (state, _storage) = event_log_state().await;
-
-        // Clone app router so we can issue multiple requests.
-        let router = app(state);
-
-        // Spawn the SSE stream reader in a background task.
-        let stream_router = router.clone();
-        let handle = tokio::spawn(async move {
-            let resp = stream_router
-                .oneshot(
-                    Request::builder()
-                        .uri("/conversations/stream")
-                        .body(Body::empty())
-                        .unwrap(),
-                )
-                .await
-                .unwrap();
-            assert_eq!(resp.status(), StatusCode::OK);
-            body_bytes(resp.into_body()).await
-        });
-
-        // Give the stream time to subscribe.
-        tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-
-        // Create a conversation via REST.
-        let create_resp = router
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/conversations")
-                    .header("Content-Type", "application/json")
-                    .body(Body::from(r#"{"title":"E2E Conv"}"#))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-        assert_eq!(create_resp.status(), StatusCode::CREATED);
-        let json = body_json(create_resp.into_body()).await;
-        let conv_id = json["id"].as_str().expect("should have an id").to_string();
-
-        // Give the stream time to forward the upserted event.
-        tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-
-        // Delete the conversation via REST.
-        let delete_resp = router
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("DELETE")
-                    .uri(format!("/conversations/{conv_id}"))
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-        assert_eq!(delete_resp.status(), StatusCode::NO_CONTENT);
-
-        // Give the stream time to forward the deleted event, then drop
-        // broadcaster indirectly by dropping the router.
-        tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-        drop(router);
-
-        let body = tokio::time::timeout(std::time::Duration::from_secs(5), handle)
-            .await
-            .expect("stream should finish within timeout")
-            .expect("stream task should not panic");
-        let text = String::from_utf8_lossy(&body);
-
-        assert!(
-            text.contains("event: upserted"),
-            "stream should contain upserted event from REST create, got: {text}"
-        );
-        assert!(
-            text.contains("E2E Conv"),
-            "upserted event should contain conversation title"
-        );
-        assert!(
-            text.contains("event: deleted"),
-            "stream should contain deleted event from REST delete, got: {text}"
-        );
-        assert!(
-            text.contains(&conv_id),
-            "deleted event should contain the conversation ID"
-        );
-    }
-}
